Split versioned models and Version concerns into separate files
