+++ lib/vestal_versions.rb
@@ -47,6 +47,8 @@ module VestalVersions
   autoload :VersionTagging
   autoload :Versioned
   autoload :Versions
+  autoload :VersionUsers
+  autoload :VersionVersionTagging
 
   class << self
     delegate :config, :configure, :to => Version
